refactor(client): Relocate ad-related components, hooks, and types to shared feature and lib directories.
