--- client/app/admin/ads/index.ts
@@ -1,8 +1,0 @@
-/**
- * What's this file?
- * 
- * This file is a barrel file for the ads module.
- * It exports all the public APIs of the ads module.
- */
-
-export { default as AdSpace } from "./components/AdSpace";
