+++ client/components/features/admin/ads/AdPlaceholder.tsx
