+++ client/lib/ads/types.ts
